+++ electron/modules/utils.ts
